more vectorizers
