+++ src/Redis.OM.Vectorizers.HuggingFace/HuggingFaceApiSentenceVectorizer.cs
@@ -1,43 +1,34 @@
-﻿using System.Net.Http.Json;
+using System.Net.Http.Json;
+using Redis.OM.Contracts;
 using Redis.OM.Modeling;
 
 namespace Redis.OM.Vectorizers.HuggingFace;
 
-public class HuggingFaceApiSentenceVectorizer : VectorizerAttribute
+public class HuggingFaceApiSentenceVectorizer : IVectorizer<string>
 {
-    public string? ModelId { get; set; }
-    public override VectorType VectorType => VectorType.FLOAT32;
-    private int? _dim;
-
-    public override int Dim
+    public HuggingFaceApiSentenceVectorizer(string authToken, string modelId, int dim)
     {
-        get
-        {
-            if (_dim is not null) return _dim.Value;
-            const string testString = "This is a vector dimensionality probing query";
-            var floats = GetFloats(testString);
-            _dim = floats.Length;
-
-            return _dim.Value;
-        }
+        _huggingFaceAuthToken = authToken;
+        ModelId = modelId;
+        Dim = dim;
     }
-
-    public override byte[] Vectorize(object obj)
+    
+    private readonly string _huggingFaceAuthToken;
+    public string ModelId { get; }
+    public VectorType VectorType => VectorType.FLOAT32;
+    
+    public int Dim { get; }
+    public byte[] Vectorize(string str)
     {
-        var s = (string)obj;
-        var floats = GetFloats(s);
-        return floats.SelectMany(BitConverter.GetBytes).ToArray();
+        return GetFloats(str).SelectMany(BitConverter.GetBytes).ToArray();
     }
-
+    
     public float[] GetFloats(string s)
     {
         var client = Configuration.Instance.Client;
-        var modelId = ModelId ?? Configuration.Instance["REDIS_OM_HF_MODEL_ID"];
-        if (modelId is null) throw new InvalidOperationException("Model Id Required to use Hugging Face API.");
-
         var requestContent = JsonContent.Create(new
         {
-            inputs = new string[] { s },
+            inputs = new [] { s },
             options = new { wait_for_model = true }
         });
 
@@ -46,10 +37,10 @@ public class HuggingFaceApiSentenceVectorizer : VectorizerAttribute
             Method = HttpMethod.Post,
             Content = requestContent,
             RequestUri =
-                new Uri($"{Configuration.Instance.HuggingFaceBaseAddress}/pipeline/feature-extraction/{modelId}"),
+                new Uri($"{Configuration.Instance.HuggingFaceBaseAddress}/pipeline/feature-extraction/{ModelId}"),
             Headers =
             {
-                { "Authorization", $"Bearer {Configuration.Instance.HuggingFaceAuthorizationToken}" }
+                { "Authorization", $"Bearer {_huggingFaceAuthToken}" }
             }
         };
