+++ src/Redis.OM/Contracts/IVectorizer.cs
@@ -5,7 +5,8 @@ namespace Redis.OM.Contracts
     /// <summary>
     /// Converter of objects into vectors.
     /// </summary>
-    public interface IVectorizer
+    /// <typeparam name="T">The type to be vectorized.</typeparam>
+    public interface IVectorizer<T>
     {
         /// <summary>
         /// Gets the vector Type generated by the vectorizer.
@@ -22,6 +23,6 @@ namespace Redis.OM.Contracts
         /// </summary>
         /// <param name="obj">the object to convert.</param>
         /// <returns>A byte array containing the vectorized data.</returns>
-        byte[] Vectorize(object obj);
+        byte[] Vectorize(T obj);
     }
 }
